5.1. Add board controller and move task rendering logics to board controller
