+++ src/controllers/board.js
@@ -1,0 +1,92 @@
+
+import SortComponent from '../components/sort';
+import TasksComponent from '../components/tasks';
+import TaskComponent from '../components/task';
+import TaskEditComponent from '../components/task-edit';
+import NoTasksComponent from '../components/no-tasks';
+import LoadMoreButtonComponent from '../components/load-more-button';
+
+import {render, remove, replace, RenderPosition} from '../utils/render';
+
+const INITIAL_TASKS_SHOWN_NUMBER = 8;
+const BY_BUTTON_TASKS_SHOWN_NUMBER = 8;
+
+const renderTask = (taskListElement, task) => {
+  const onEscKeyDown = (evt) => {
+    const isEscKeyDown = evt.key === `Escape` || evt.key === `Esc`;
+
+    if (isEscKeyDown) {
+      replaceEditToTask();
+      document.removeEventListener(`keydown`, onEscKeyDown);
+    }
+  };
+
+  const replaceEditToTask = () => {
+    replace(taskComponent, taskEditComponent);
+  };
+
+  const replaceTaskToEdit = () => {
+    replace(taskEditComponent, taskComponent);
+  };
+
+  const taskComponent = new TaskComponent(task);
+  taskComponent.setEditButtonClickHandler(() => {
+    replaceTaskToEdit();
+    document.addEventListener(`keydown`, onEscKeyDown);
+  });
+
+  const taskEditComponent = new TaskEditComponent(task);
+  taskEditComponent.setSubmitHandler(replaceEditToTask);
+
+  render(taskListElement, taskComponent, RenderPosition.BEFOREEND);
+};
+
+export default class BoardController {
+  constructor(container) {
+    this._container = container;
+
+    this._noTasksComponent = new NoTasksComponent();
+    this._sortComponent = new SortComponent();
+    this._tasksComponent = new TasksComponent();
+    this._loadMoreButtonComponent = new LoadMoreButtonComponent();
+  }
+
+  render(tasks) {
+    const container = this._container.getElement();
+    const isAllTasksArchived = tasks.every((task) => task.isArchive);
+
+    if (isAllTasksArchived || tasks.length === 0) {
+      render(container, this._noTasksComponent, RenderPosition.BEFOREEND);
+      return;
+    }
+
+    render(container, this._sortComponent, RenderPosition.BEFOREEND);
+    render(container, this._tasksComponent, RenderPosition.BEFOREEND);
+
+    const taskListElement = this._tasksComponent.getElement();
+
+    // Render tasks
+    let shownTasksNumber = INITIAL_TASKS_SHOWN_NUMBER;
+    tasks.slice(0, shownTasksNumber)
+      .forEach((task) => {
+        renderTask(taskListElement, task);
+      });
+
+    // Render LOAD MORE button
+    render(container, this._loadMoreButtonComponent, RenderPosition.BEFOREEND);
+
+    // LOAD MORE button logic
+    this._loadMoreButtonComponent.setClickHandler(() => {
+      const prevShownTasksNumber = shownTasksNumber;
+      shownTasksNumber += BY_BUTTON_TASKS_SHOWN_NUMBER;
+
+      tasks.slice(prevShownTasksNumber, shownTasksNumber).forEach((task) => {
+        renderTask(taskListElement, task);
+      });
+
+      if (shownTasksNumber >= tasks.length) {
+        remove(this._loadMoreButtonComponent);
+      }
+    });
+  }
+}
